PDFtoXMLUsingExcel/legacy/requirements_standalone.txt
# Standalone PDF to DOCX/XML Converter Requirements
# Install with: pip install -r requirements_standalone.txt

# PDF Processing
pdf2docx==0.5.8
PyMuPDF==1.26.4
pdf2image==1.17.0
pymupdf==1.26.4

# Image Processing
Pillow==11.3.0

# OpenAI API for AI-powered conversion
openai==1.109.1

# Document Conversion
pypandoc==1.14
python-docx==1.1.2

# XML Processing
lxml==6.0.1

# Async support
aiofiles==24.1.0

# CLI
argparse  # Built-in, but listed for clarity

# Note: System dependencies required:
# - Poppler (for pdf2image):
#     Ubuntu/Debian: sudo apt-get install poppler-utils
#     macOS: brew install poppler
#     Windows: Download from https://github.com/osber/poppler-windows
#
# - Pandoc (for pypandoc):
#     Ubuntu/Debian: sudo apt-get install pandoc
#     macOS: brew install pandoc
#     Windows: Download from https://pandoc.org/installing.html
